- BaseType's Size() is optimized by accessing array instead of switch-case. If invalid, it returns zero instead of 255, reverted since we now have Valid method and if invalid it will be skipped in decoder and encoder to avoid integer divide by zero.
- Proto Value's `Valid()` and `MarshalAppend` is optimized for single value, by accessing `v.num` and `v.any` directly instead of via method. `Type()` is optimized by removing if-statement before type's switch-case.
- Encoder's Validator now will only `memmove` slice's value when `i != valid`.

### Benchmark
#### Decoder
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                    │    old.txt     │               new.txt                │
                    │     sec/op     │    sec/op     vs base                │
DecodeMessageData-4    3.748µ ±   3%   3.538µ ± 11%   -5.62% (p=0.019 n=10)
Decode-4              137.54m ±  49%   95.93m ±  5%  -30.25% (p=0.000 n=10)
DecodeWithFiledef-4   119.82m ± 105%   91.39m ± 31%  -23.73% (p=0.005 n=10)
geomean                3.953m          3.142m        -20.52%

                    │   old.txt    │               new.txt               │
                    │     B/op     │     B/op      vs base               │
DecodeMessageData-4     144.0 ± 1%     144.0 ± 1%       ~ (p=1.000 n=10)
Decode-4              73.49Mi ± 0%   73.49Mi ± 0%       ~ (p=0.157 n=10)
DecodeWithFiledef-4   49.70Mi ± 0%   49.70Mi ± 0%  -0.00% (p=0.002 n=10)
geomean               813.6Ki        813.6Ki       -0.00%

                    │   old.txt   │               new.txt                │
                    │  allocs/op  │  allocs/op   vs base                 │
DecodeMessageData-4    11.00 ± 0%    11.00 ± 0%       ~ (p=1.000 n=10) ¹
Decode-4              100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=10) ¹
DecodeWithFiledef-4   100.1k ± 0%   100.1k ± 0%  -0.00% (p=0.000 n=10)
geomean               4.793k        4.793k       -0.00%
¹ all samples are equal

```

#### Encoder
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/encoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                                             │   old.txt    │               new.txt                │
                                             │    sec/op    │    sec/op     vs base                │
Encode/normal_header_zero-4                    88.25m ±  3%   67.01m ±  2%  -24.07% (p=0.000 n=10)
Encode/normal_header_15-4                      87.50m ±  3%   69.92m ± 58%  -20.08% (p=0.023 n=10)
Encode/compressed_timestamp_header-4           76.55m ±  2%   88.22m ± 33%  +15.24% (p=0.002 n=10)
EncodeWriterAt/normal_header_zero-4            41.26m ±  2%   39.86m ± 47%        ~ (p=0.796 n=10)
EncodeWriterAt/normal_header_15-4              41.79m ± 27%   35.82m ± 30%  -14.27% (p=0.029 n=10)
EncodeWriterAt/compressed_timestamp_header-4   43.32m ± 34%   32.95m ±  2%  -23.94% (p=0.000 n=10)
EncodeFull-4                                   76.03m ± 32%   63.36m ±  7%  -16.67% (p=0.000 n=10)
geomean                                        61.58m         53.36m        -13.34%

                                             │  old.txt   │               new.txt               │
                                             │    B/op    │    B/op     vs base                 │
Encode/normal_header_zero-4                    183.0 ± 1%   181.0 ± 0%  -1.09% (p=0.000 n=10)
Encode/normal_header_15-4                      455.0 ± 0%   453.0 ± 0%  -0.44% (p=0.000 n=10)
Encode/compressed_timestamp_header-4           181.0 ± 1%   180.0 ± 1%       ~ (p=0.338 n=10)
EncodeWriterAt/normal_header_zero-4            91.00 ± 0%   90.00 ± 1%  -1.10% (p=0.011 n=10)
EncodeWriterAt/normal_header_15-4              227.0 ± 0%   226.0 ± 0%  -0.44% (p=0.000 n=10)
EncodeWriterAt/compressed_timestamp_header-4   90.00 ± 0%   90.00 ± 0%       ~ (p=1.000 n=10) ¹
EncodeFull-4                                   94.00 ± 1%   92.00 ± 1%  -2.13% (p=0.000 n=10)
geomean                                        159.6        158.3       -0.82%
¹ all samples are equal

                                             │  old.txt   │               new.txt               │
                                             │ allocs/op  │ allocs/op   vs base                 │
Encode/normal_header_zero-4                    4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
Encode/normal_header_15-4                      14.00 ± 0%   14.00 ± 0%       ~ (p=1.000 n=10) ¹
Encode/compressed_timestamp_header-4           4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
EncodeWriterAt/normal_header_zero-4            2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
EncodeWriterAt/normal_header_15-4              7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=10) ¹
EncodeWriterAt/compressed_timestamp_header-4   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
EncodeFull-4                                   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                        3.850        3.850       +0.00%
¹ all samples are equal

```